Create new monitor layout system
Previously, I had to keep rerunning a config whenever I reloaded herbs
because it kept breaking the scrrens layout. Now every script that
changes the layout stores a key in a hidden file in the home directory.
Herbstluftwm's config is aware of this file and sets the screen layout
approppriately.